+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [v0.1.9] - 2025-06-24
+### Fixed
+- Fix error caused by nested arrays in pipeline
+
+### Changed
+- Body data can now be provided as multi-value array keys
+
 ## [v0.1.8] - 2025-06-23
 ### Fixed
 - Ensure new lines for multiple body options in curl output